🚨 fix: lint — function/identifier/file-length over caps
Three swiftlint --strict failures introduced earlier in this branch:

- ScenarioEditorViewModel.save: 51-line body over the 50-line cap.
  Extract `runCommitTimeValidation(_:)` private helper carrying the
  validateForCommit do-catch.
- ResultMarkdownExporter.formatOutput: `pt` identifier under the
  3-char minimum. Rename to `resolved`.
- PhaseEditorSheet: 428 lines over the 400-line cap. Move
  `outputFieldsSection` + the new `canonicalFieldHint` into a sibling
  `PhaseEditorSheet+CanonicalFieldHint.swift` (matches the existing
  `+ConditionalSection` / `+EventInjectSection` extension pattern).
  Drop `private` on `@State newOutputFieldName` so the sibling
  extension can bind it.
